Move ui components into folder
