+++ components/ui/PrimaryButton.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Pressable, StyleSheet, Text, View } from 'react-native';
 
-import Colors from '../constants/colors';
+import Colors from '../../constants/colors';
 
 const PrimaryButton = ({ children, onPress }) => {
   return (
